Change size field type to int to avoid type conversions.
The size field in the code base currently uses the byte type. However,
this can lead to unnecessary type conversions between int and byte
throughout the code. This commit changes the size field type to int to
avoid these conversions

Bug: 293811969
Test: TH

+++ src/android/net/apf/ApfGenerator.java
@@ -146,7 +146,7 @@ public class ApfGenerator {
         public final IntImmediateType mImmediateType;
         // TODO: remove mSigned and mImmSize fields in follow up CL
         public final boolean mSigned;
-        public final byte mImmSize;
+        public final int mImmSize;
         public final int mValue;
 
         IntImmediate(int value, boolean signed) {
@@ -154,7 +154,7 @@ public class ApfGenerator {
         }
 
         // TODO: Remove this constructor in follow up CLs.
-        IntImmediate(int value, boolean signed, byte size) {
+        IntImmediate(int value, boolean signed, int size) {
             mValue = value;
             mSigned = signed;
             mImmSize = size;
@@ -236,7 +236,7 @@ public class ApfGenerator {
         private final byte mRegister; // A "Register" value.
         public final List<IntImmediate> mIntImms = new ArrayList<>();
         // When mOpcode is a jump:
-        private byte mTargetLabelSize;
+        private int mTargetLabelSize;
         private String mTargetLabel;
         // When mOpcode == Opcodes.LABEL:
         private String mLabel;
@@ -300,7 +300,7 @@ public class ApfGenerator {
                 return 0;
             }
             int size = 1;
-            byte maxImmSize = getMaxImmSize();
+            int maxImmSize = getMaxImmSize();
             // For the copy opcode, the last imm is the length field is always 1 byte
             size += mIntImms.size() * maxImmSize;
             if (mTargetLabel != null) {
@@ -332,8 +332,8 @@ public class ApfGenerator {
         /**
          * Assemble value for instruction size field.
          */
-        private byte generateImmSizeField() {
-            byte immSize = getMaxImmSize();
+        private int generateImmSizeField() {
+            int immSize = getMaxImmSize();
             // Encode size field to fit in 2 bits: 0->0, 1->1, 2->2, 3->4.
             return immSize == 4 ? 3 : immSize;
         }
@@ -342,7 +342,7 @@ public class ApfGenerator {
          * Assemble first byte of generated instruction.
          */
         private byte generateInstructionByte() {
-            byte sizeField = generateImmSizeField();
+            int sizeField = generateImmSizeField();
             return (byte)((mOpcode << 3) | (sizeField << 1) | mRegister);
         }
 
@@ -355,7 +355,7 @@ public class ApfGenerator {
          * be sign extended and the truncation should simply throw away their signed
          * upper bits.
          */
-        private int writeValue(int value, byte[] bytecode, int writingOffset, byte immSize) {
+        private int writeValue(int value, byte[] bytecode, int writingOffset, int immSize) {
             for (int i = immSize - 1; i >= 0; i--) {
                 bytecode[writingOffset++] = (byte)((value >> (i * 8)) & 255);
             }
@@ -371,7 +371,7 @@ public class ApfGenerator {
             }
             int writingOffset = offset;
             bytecode[writingOffset++] = generateInstructionByte();
-            byte maxImmSize = getMaxImmSize();
+            int maxImmSize = getMaxImmSize();
             if (mTargetLabel != null) {
                 writingOffset = writeValue(calculateTargetLabelOffset(), bytecode, writingOffset,
                         maxImmSize);
@@ -397,10 +397,10 @@ public class ApfGenerator {
          * byte, hence why this function simply takes the maximum of those sizes, so neither is
          * truncated.
          */
-        private byte getMaxImmSize() {
-            byte maxSize = mTargetLabelSize;
+        private int getMaxImmSize() {
+            int maxSize = mTargetLabelSize;
             for (int i = 0; i < mIntImms.size(); ++i) {
-                maxSize = (byte) Math.max(maxSize, mIntImms.get(i).mImmSize);
+                maxSize = Math.max(maxSize, mIntImms.get(i).mImmSize);
             }
             return maxSize;
         }
@@ -1266,7 +1266,7 @@ public class ApfGenerator {
     /**
      * Calculate the size of the imm.
      */
-    private static byte calculateImmSize(int imm, boolean signed) {
+    private static int calculateImmSize(int imm, boolean signed) {
         if (imm == 0) {
             return 0;
         }
